Changes:- 1) Added Two extra plugs and made it generic so that it can be used for other controllers. 2) Modified user_controller to accomodate recent changes 3) Removed warnings
