+++ apps/acqdat_api/lib/acqdat_api_web/plugs/load_org.ex
@@ -1,0 +1,25 @@
+defmodule AcqdatApiWeb.Plug.LoadOrg do
+  import Plug.Conn
+  alias AcqdatCore.Model.Organisation, as: OrgModel
+
+  @spec init(any) :: any
+  def init(default), do: default
+
+  @spec call(Plug.Conn.t(), any) :: Plug.Conn.t()
+  def call(%{params: %{"org_id" => org_id}} = conn, _params) do
+    check_org(conn, org_id)
+  end
+
+  defp check_org(conn, org_id) do
+    {org_id, _} = Integer.parse(org_id)
+
+    case OrgModel.get(org_id) do
+      {:ok, org} ->
+        assign(conn, :org, org)
+
+      {:error, _message} ->
+        conn
+        |> put_status(404)
+    end
+  end
+end
