+++ apps/acqdat_core/test/support/factory/factory.ex
@@ -20,7 +20,6 @@ defmodule AcqdatCore.Support.Factory do
   alias AcqdatCore.Schema.{
     User,
     UserSetting,
-    Device,
     Sensor,
     DigitalTwin,
     Organisation,
